# Rewrite burn_rates_client using @elastic/elasticsearch-mock