--- a/package.json
+++ b/package.json
@@ -1457,6 +1457,7 @@
     "@cypress/debugging-proxy": "2.0.1",
     "@cypress/grep": "^4.0.1",
     "@cypress/webpack-preprocessor": "^6.0.2",
+    "@elastic/elasticsearch-mock": "^2.1.0",
     "@elastic/eslint-plugin-eui": "2.4.0",
     "@elastic/makelogs": "^6.1.1",
     "@elastic/synthetics": "^1.19.0",