Related to https://github.com/elastic/kibana/issues/234366

## Summary

Command

```
node ./scripts/jest --no-cache --config="x-pack/solutions/observability/plugins/slo/jest.config.js" --runInBand --coverage=false --passWithNoTests --silent x-pack/solutions/observability/plugins/slo/server/services/burn_rates_client.test.ts 
```
### Improvement

|Test|Before|After|Difference|
|---|---|---|---|
|burn_rates_client.test.ts|5.467 s|1.631 s| +70.2% faster|

### Before
<img width="2852" height="646" alt="image" src="https://github.com/user-attachments/assets/1dd7d6ce-b639-4f23-8a22-016287da40eb" />


### After
<img width="2866" height="646" alt="image" src="https://github.com/user-attachments/assets/ee228706-bbbd-4407-9cdf-62286500c0a1" />